pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkbay-auth"
version = "0.1.0"
description = "Plugin di autenticazione completo per FastAPI con JWT, refresh token e recupero password"
authors = [
    {name = "LinkBay CMS Team"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["fastapi", "authentication", "jwt", "oauth2", "auth"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
  "fastapi>=0.104.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic[email]>=2.0.0",
  "python-multipart>=0.0.6"  # Per OAuth2PasswordRequestForm
]

[project.optional-dependencies]
dev = [
  "uvicorn[standard]>=0.24.0",
  "pytest>=7.4.0",
  "httpx>=0.25.0"
]

[project.urls]
Homepage = "https://github.com/linkbay/linkbay-auth"
Documentation = "https://github.com/linkbay/linkbay-auth#readme"
Repository = "https://github.com/linkbay/linkbay-auth"

[tool.setuptools]
packages = ["linkbay-auth"]
